Back4app

+++ COPDAsthma/app/src/main/java/com/example/copd_asthma/data/Users.kt
@@ -3,24 +3,24 @@ package com.example.copd_asthma.data
 
 
 
-data class Users (
-    val name : String,
-    val email : String,
-    val password : String,
+data class Users(
+    val name: String,
+    val email: String,
+    val password: String,
     val age: Int,
     val packHistory: Int,
-    val fev1: Int,
-    val fvc: Int,
-    val mwt1: Int,
-    val mwt2: Int,
-    val mwtBest: Int = maxOf(mwt1, mwt2),
+    val fev1: Double,
+    val fvc: Double,
+    val mwt1: Double,
+    val mwt2: Double,
+    val mwtBest: Double = maxOf(mwt1, mwt2),
     val severity: String = calculateSeverity(fev1),
     val gender: String,
     val diabetes: String,
     val hypertension: String
 ) {
     companion object {
-        private fun calculateSeverity(fev1: Int): String {
+        private fun calculateSeverity(fev1: Double): String {
             val thresholds = listOf(80, 50, 30, 0)
             val labels = listOf("Mild", "Moderate", "Severe", "Very Severe")
